requirements.txt
# Automated Git Commit Message Generator - Dependencies
# Python 3.8+

# Deep Learning
torch>=2.0.0
transformers>=4.30.0

# Tokenization
sentencepiece>=0.1.99

# Git / Data
GitPython>=3.1.40
tqdm>=4.65.0
datasets>=2.14.0

# Evaluation
nltk>=3.8.0
sacrebleu>=2.3.0
rouge-score>=0.1.2

# Utils
numpy>=1.24.0
pyyaml>=6.0
